backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
openai
pydantic
